Added connectivity to driver station GUI
Added logic in WPIRBMain initialization to always connect to the driver
station interface. This connection is used to determine the active FCS
mode. Request and response logic for the mode is added into the main
loop. The driver station is currently assumed to be at a fixed host and
port.